pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["gateway", "api_gateway_project"]

[project]
name = "django-api-gateway"
version = "0.1.2"
description = "API Gateway with Django and IFTTT-like interface"
readme = "README.md"
authors = [
    { name = "Peter Rosemann", email = "dkdndes@gmail.com" },
    { name = "OpenHands",     email = "openhands@all-hands.dev" }
]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
]
requires-python = ">=3.10"
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.14.0",
    "requests>=2.28.0",
    "coreapi>=2.3.3",
    "pyyaml>=6.0",
    "django-htmx>=1.23.0",
    "django-cors-headers>=4.7.0",
    "django-allauth>=65.8.0",
    "jwt>=1.3.1",
    "pytest>=8.3.5",
    "certifi>=2025.4.26",
]

keywords = ["django", "api-gateway", "ifttt", "webhooks"]

[project.urls]
Homepage     = "https://github.com/dkdndes/django-api-gateway"
"Bug-Tracker"  = "https://github.com/dkdndes/django-api-gateway/issues"

[project.optional-dependencies]
test = ["pytest", "pytest-django", "pytest-cov"]
docs = ["mkdocs", "mkdocstrings"]
dev = ["pytest", "pytest-django", "pytest-cov", "mkdocs", "mkdocstrings"]
